requirements.txt
streamlit
tensorflow
numpy
Pillow
pickle-mixin